test mapping name
