+++ pom.xml
@@ -142,6 +142,10 @@
     <build>
         <plugins>
 
+            <plugin>
+                <artifactId>maven-dependency-plugin</artifactId>
+                <version>3.1.2</version>
+            </plugin>
             <!-- COMPILER LOGIC -->
 
             <plugin>
