+++ src/main/scala/org/finos/legend/spark/LegendUtils.scala
@@ -1,9 +1,26 @@
 package org.finos.legend.spark
 
-import org.apache.spark.sql.types.DataType
+import java.io.StringReader
+import java.util.Collections
+
+import net.sf.jsqlparser.parser.CCJSqlParserManager
+import net.sf.jsqlparser.statement.select.{PlainSelect, Select}
+import org.apache.spark.sql.types.{BinaryType, BooleanType, DataType, DateType, DoubleType, FloatType, IntegerType, LongType, StringType, TimestampType}
+import org.finos.legend.engine.language.pure.compiler.toPureGraph.{HelperValueSpecificationBuilder, PureModel}
+import org.finos.legend.engine.language.pure.grammar.from.PureGrammarParser
+import org.finos.legend.engine.plan.generation.PlanGenerator
+import org.finos.legend.engine.plan.generation.transformers.LegendPlanTransformers
+import org.finos.legend.engine.plan.platform.PlanPlatform
+import org.finos.legend.engine.protocol.pure.v1.model.executionPlan.SingleExecutionPlan
+import org.finos.legend.engine.protocol.pure.v1.model.executionPlan.nodes.SQLExecutionNode
 import org.finos.legend.engine.protocol.pure.v1.model.packageableElement.domain._
+import org.finos.legend.engine.protocol.pure.v1.model.valueSpecification.ValueSpecification
+import org.finos.legend.engine.protocol.pure.v1.model.valueSpecification.raw.Lambda
+import org.finos.legend.pure.generated.core_relational_relational_router_router_extension.Root_meta_pure_router_extension_defaultRelationalExtensions__RouterExtension_MANY_
 import org.finos.legend.pure.generated.{Root_meta_relational_mapping_RelationalPropertyMapping_Impl, Root_meta_relational_mapping_RootRelationalInstanceSetImplementation_Impl, Root_meta_relational_metamodel_TableAliasColumn_Impl, Root_meta_relational_metamodel_relation_Table_Impl}
 import org.finos.legend.pure.m3.coreinstance.meta.pure.mapping.Mapping
+import org.finos.legend.pure.m3.coreinstance.meta.pure.metamodel.function.LambdaFunction
+import org.finos.legend.pure.m3.coreinstance.meta.pure.runtime
 import org.finos.legend.sdlc.domain.model.entity.Entity
 
 import scala.collection.JavaConverters._
@@ -45,6 +62,98 @@ object LegendUtils {
   }
 
   /**
+   * A complexity when dealing with nested fields is to make sure we call a field with a [parent.child] syntax
+   *
+   * @param fieldName       the name of the field
+   * @param parentFieldName the name of the parent field this field is included into (empty for top level object)
+   * @return the concatenation of [parent.child] to reference this field
+   */
+  def childFieldName(fieldName: String, parentFieldName: String): String =
+    if (parentFieldName.isEmpty) s"`$fieldName`" else s"$parentFieldName.`$fieldName`"
+
+  /**
+   * We need to ensure fields (especially nested) are enclosed with backticks for spark SQL
+   *
+   * @param name of the field
+   * @return cleaned version of the field to query
+   */
+  def cleanColumnName(name: String): String = name.replaceAll("`", "")
+
+  /**
+   * Simple mapping function that converts Legend data type into Spark SQL DataType
+   *
+   * @return the corresponding DataType for a given legend type
+   */
+  def convertDataTypeFromString(returnType: String): DataType = {
+    returnType match {
+      case "String" => StringType
+      case "Boolean" => BooleanType
+      case "Binary" => BinaryType
+      case "Integer" => IntegerType
+      case "Number" => LongType
+      case "Float" => FloatType
+      case "Decimal" => DoubleType
+      case "Date" => DateType
+      case "StrictDate" => DateType
+      case "DateTime" => TimestampType
+      case _ =>
+        throw new IllegalArgumentException(s"entity of class [${returnType}] is not supported as primitive")
+    }
+  }
+
+  /**
+   * Parse SQL to retrieve WHERE clause and table alias
+   * TODO: find a way not to generate the full SQL but only visit the lambda condition
+   *
+   * @param executionPlan generated SQL plan from legend engine
+   * @return the WHERE clause of the generated SQL expression
+   */
+  def parseSql(executionPlan: SQLExecutionNode): String = {
+    val parserRealSql = new CCJSqlParserManager()
+    val select = parserRealSql.parse(new StringReader(executionPlan.sqlQuery)).asInstanceOf[Select].getSelectBody.asInstanceOf[PlainSelect]
+    val alias = s"${select.getFromItem.getAlias.getName}."
+    val where = select.getWhere
+    where.toString.replaceAll(alias, "")
+  }
+
+  def generateExecutionPlan(query: String, legendMapping: Mapping, legendRuntime: runtime.Runtime, pureModel: PureModel): SingleExecutionPlan = {
+    PlanGenerator.generateExecutionPlan(
+      buildLambda(query, pureModel),
+      legendMapping,
+      legendRuntime,
+      null,
+      pureModel,
+      "vX_X_X",
+      PlanPlatform.JAVA,
+      "1.0",
+      Root_meta_pure_router_extension_defaultRelationalExtensions__RouterExtension_MANY_(pureModel.getExecutionSupport),
+      LegendPlanTransformers.transformers
+    )
+  }
+
+  /**
+   * Build the value specification for a lambda function
+   *
+   * @param lambdaString the string representation of the lambda function
+   * @param pureModel    the pure model
+   * @return the compiled function
+   */
+  def buildLambda(lambdaString: String, pureModel: PureModel): LambdaFunction[_] = {
+    val function = buildLambda(lambdaString)
+    val lambda = new Lambda()
+    lambda.body = Collections.singletonList(function)
+    HelperValueSpecificationBuilder.buildLambda(lambda, pureModel.getContext)
+  }
+
+  def buildLambda(lambdaString: String): ValueSpecification = {
+    val parser = PureGrammarParser.newInstance()
+    val parsed = parser.parseLambda(lambdaString, "id", true)
+    require(parsed.body != null && parsed.body.size() > 0)
+    parsed.body.get(0)
+  }
+
+
+  /**
    * Utility class to manipulate Legend Property object (i.e. a Field) easily
    *
    * @param property the Legend Property object
@@ -77,7 +186,7 @@ object LegendUtils {
      *
      * @return the corresponding DataType for a given legend type
      */
-    def convertDataType: DataType = Legend.convertDataTypeFromString(property.`type`)
+    def convertDataType: DataType = convertDataTypeFromString(property.`type`)
   }
 
   implicit class ClassImpl(clazz: Class) {
@@ -111,7 +220,6 @@ object LegendUtils {
     /**
      * Convert a JSON constraint into its lambda definition
      * Constraints are defined as anonymous lambda function incompatible with Pure execution plan
-     * We fix that with a simple hack for now //TODO: fix this hack
      *
      * @return the Lambda representation of a constraint
      */
@@ -122,20 +230,19 @@ object LegendUtils {
         .dropRight(1)
         .drop(2)
         .replaceAll("\\\\n", "")
-        .replaceAll("\\s+", "")
     }
   }
 
 
   implicit class TransformationImpl(transformation: Root_meta_relational_mapping_RootRelationalInstanceSetImplementation_Impl) {
 
-    def getTransformations: Seq[WithColumnRenamed] = {
+    def getTransformations: Seq[Transformation] = {
       transformation._propertyMappings.asScala.flatMap({ o =>
         o match {
           case p: Root_meta_relational_mapping_RelationalPropertyMapping_Impl =>
             p._relationalOperationElement match {
               case e: Root_meta_relational_metamodel_TableAliasColumn_Impl =>
-                Some(WithColumnRenamed(p._property()._name(), e._columnName()))
+                Some(Transformation(p._property()._name(), e._columnName()))
               case _ => None
             }
           case _ =>
